requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
python-dotenv>=1.0
jupyter>=1.0
matplotlib>=3.8
seaborn>=0.13
nbformat>=5.0
